requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   groq
    #   httpx
attrs==25.4.0
    # via aiohttp
cachetools==6.2.2
    # via google-auth
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via
    #   pdfminer-six
    #   requests
cryptography==46.0.3
    # via pdfminer-six
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via groq
filetype==1.2.0
    # via langchain-google-genai
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
google-ai-generativelanguage==0.9.0
    # via langchain-google-genai
google-api-core==2.28.1
    # via google-ai-generativelanguage
google-auth==2.43.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
groq==0.36.0
    # via langchain-groq
grpcio==1.76.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   grpcio-status
grpcio-status==1.76.0
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   groq
    #   langgraph-sdk
    #   langsmith
httpx-sse==0.4.3
    # via langchain-community
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.1.0
    # via employeerecuiteragent (pyproject.toml)
langchain-classic==1.0.0
    # via langchain-community
langchain-community==0.4.1
    # via employeerecuiteragent (pyproject.toml)
langchain-core==1.1.0
    # via
    #   langchain
    #   langchain-classic
    #   langchain-community
    #   langchain-google-genai
    #   langchain-groq
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-google-genai==3.2.0
    # via employeerecuiteragent (pyproject.toml)
langchain-groq==1.1.0
    # via employeerecuiteragent (pyproject.toml)
langchain-text-splitters==1.0.0
    # via langchain-classic
langgraph==1.0.4
    # via
    #   employeerecuiteragent (pyproject.toml)
    #   langchain
langgraph-checkpoint==3.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.5
    # via langgraph
langgraph-sdk==0.2.10
    # via langgraph
langsmith==0.4.49
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.3.5
    # via langchain-community
orjson==3.11.4
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
pdfminer-six==20251107
    # via employeerecuiteragent (pyproject.toml)
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==6.33.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   groq
    #   langchain
    #   langchain-classic
    #   langchain-core
    #   langchain-google-genai
    #   langgraph
    #   langsmith
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via langchain-community
python-dotenv==1.2.1
    # via pydantic-settings
pyyaml==6.0.3
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
requests==2.32.5
    # via
    #   google-api-core
    #   langchain-classic
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via groq
sqlalchemy==2.0.44
    # via
    #   langchain-classic
    #   langchain-community
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
typing-extensions==4.15.0
    # via
    #   groq
    #   grpcio
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
xxhash==3.6.0
    # via langgraph
yarl==1.22.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
